+++ src/components/YoutubeEmbed/index.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import * as s from './style.module.less';
+
+export function YoutubeEmbed({ url }) {
+  return (
+    <div className={s.youtubeContainer}>
+      <iframe
+        title="Youtube Embed"
+        width="560"
+        height="315"
+        src={`https://www.youtube-nocookie.com/embed/${url}?rel=0`}
+        frameBorder="0"
+        allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+        allowFullScreen
+        loading="lazy"
+      ></iframe>
+    </div>
+  );
+}
+
+// Default export needed for lazy loading
+export default YoutubeEmbed;
