+++ src/components/YoutubeEmbed/style.module.less
@@ -1,0 +1,17 @@
+@import '../style/vars.less';
+
+.youtubeContainer {
+  position: relative;
+  padding-bottom: 56.25%;
+  margin-bottom: 2rem;
+  .contentGridSlim();
+  pointer-events: all;
+
+  iframe {
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+  }
+}
